fix: fix overlays not being shown and annotatedRenderOutput
The method to build overlays was present but never used, therefore no delete icon or loading indicator have been drawn. The issues was fixed by wrapping the _buildInner output in the default build method using another stack, that appends the overlaying widgets after the built layout. The data for annotatedRenderOutput also was build, but not assigned to the instance field, which has been fixed.

Refs: #1